requirements.txt
discord.py
python-dotenv
aiosqlite
openai
asyncio
numpy
aiohttp